Exibindo estado da aplicação

+++ Api/build/Rotas/Gerenciamento/Editar.js
@@ -3,6 +3,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var Routes_1 = require("../Routes");
 var Mongo_1 = require("../../MongoDB/Mongo");
 var MongoCollections_1 = require("../../MongoDB/MongoCollections");
+var usuarios_service_1 = require("../Usuario/usuarios.service");
 var multer = require('multer');
 var express = require("express");
 var ObjectId = require('mongodb').ObjectID;
@@ -24,21 +25,24 @@ var storage = multer.diskStorage({
 var upload = multer({ storage: storage });
 // [ROUTES]----------------------------------------------------------------------------------------------
 app.put(Routes_1.Rotas.Cardapios, upload.single('file'), function (req, res) {
-    console.log(req.body.Cardapio.Src);
+    console.log("token", req.body.token);
     try {
-        var query = {
-            Dia: req.body.Cardapio.Dia,
-            Nome: req.body.Cardapio.Nome,
-            Ingredientes: req.body.Cardapio.Ingredientes,
-            Tipo: req.body.Cardapio.Tipo,
-            Src: req.body.Cardapio.Src,
-            SrcType: req.body.Cardapio.SrcType
-        };
-        console.log("gerenciamento/" + Routes_1.Rotas.Cardapios, query);
-        Mongo_1.Mongo.Edit(MongoCollections_1.Collections.Cardapio.NomeID, req.body.Cardapio._id, query).then(function (x) {
-            redisConfig_1.redisConfig.flushAll();
-            console.log(x);
-            res.send(x);
+        usuarios_service_1.UsuarioService.getByToken(req.body.token).then(function (user) {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                var query = {
+                    Dia: req.body.item.Cardapio.Dia,
+                    Nome: req.body.item.Cardapio.Nome,
+                    Ingredientes: req.body.item.Cardapio.Ingredientes,
+                    Tipo: req.body.item.Cardapio.Tipo,
+                    Src: req.body.item.Cardapio.Src,
+                    SrcType: req.body.item.Cardapio.SrcType
+                };
+                Mongo_1.Mongo.Edit(MongoCollections_1.Collections.Cardapio.NomeID, req.body.item.Cardapio._id, query).then(function (x) {
+                    redisConfig_1.redisConfig.flushAll();
+                    res.send(x);
+                });
+            }
         });
     }
     catch (err) {
@@ -47,16 +51,21 @@ app.put(Routes_1.Rotas.Cardapios, upload.single('file'), function (req, res) {
 });
 app.put(Routes_1.Rotas.InfoContato, function (req, res) {
     try {
-        var query = {
-            Telefone: req.body.InformacoesContato.Telefone,
-            Email: req.body.InformacoesContato.Email,
-            HorarioAtendimento: req.body.InformacoesContato.HorarioAtendimento,
-            Whatsapp: req.body.InformacoesContato.Whatsapp,
-            Instagram: req.body.InformacoesContato.Instagram
-        };
-        Mongo_1.Mongo.Edit(MongoCollections_1.Collections.InformacoesContato.NomeID, req.body.InformacoesContato._id, query).then(function (x) {
-            redisConfig_1.redisConfig.flushAll();
-            res.send(x);
+        usuarios_service_1.UsuarioService.getByToken(req.body.token).then(function (user) {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                var query = {
+                    Telefone: req.body.item.InformacoesContato.Telefone,
+                    Email: req.body.item.InformacoesContato.Email,
+                    HorarioAtendimento: req.body.item.InformacoesContato.HorarioAtendimento,
+                    Whatsapp: req.body.item.InformacoesContato.Whatsapp,
+                    Instagram: req.body.item.InformacoesContato.Instagram
+                };
+                Mongo_1.Mongo.Edit(MongoCollections_1.Collections.InformacoesContato.NomeID, req.body.item.InformacoesContato._id, query).then(function (x) {
+                    redisConfig_1.redisConfig.flushAll();
+                    res.send(x);
+                });
+            }
         });
     }
     catch (err) {
@@ -65,16 +74,21 @@ app.put(Routes_1.Rotas.InfoContato, function (req, res) {
 });
 app.put(Routes_1.Rotas.Sobre, function (req, res) {
     try {
-        var query = {
-            Descricao: req.body.Sobre.Descricao,
-            Nome: req.body.Sobre.Nome,
-            Servico: req.body.Sobre.Servico,
-            Historia: req.body.Sobre.Historia,
-            Slogan: req.body.Sobre.Slogan,
-        };
-        Mongo_1.Mongo.Edit(MongoCollections_1.Collections.Sobre.NomeID, req.body.Sobre._id, query).then(function (x) {
-            redisConfig_1.redisConfig.flushAll();
-            res.send(x);
+        usuarios_service_1.UsuarioService.getByToken(req.body.token).then(function (user) {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                var query = {
+                    Descricao: req.body.item.Sobre.Descricao,
+                    Nome: req.body.item.Sobre.Nome,
+                    Servico: req.body.item.Sobre.Servico,
+                    Historia: req.body.item.Sobre.Historia,
+                    Slogan: req.body.item.Sobre.Slogan,
+                };
+                Mongo_1.Mongo.Edit(MongoCollections_1.Collections.Sobre.NomeID, req.body.item.Sobre._id, query).then(function (x) {
+                    redisConfig_1.redisConfig.flushAll();
+                    res.send(x);
+                });
+            }
         });
     }
     catch (err) {
@@ -83,11 +97,18 @@ app.put(Routes_1.Rotas.Sobre, function (req, res) {
 });
 app.put(Routes_1.Rotas.PrecoMarmitex, function (req, res) {
     try {
-        var query = { "Pequena": req.body.PrecoMarmitex.Pequena,
-            "Grande": req.body.PrecoMarmitex.Grande };
-        Mongo_1.Mongo.Edit(MongoCollections_1.Collections.PrecoMarmitex.NomeID, req.body.PrecoMarmitex._id, query).then(function (x) {
-            redisConfig_1.redisConfig.flushAll();
-            res.send(x);
+        usuarios_service_1.UsuarioService.getByToken(req.body.token).then(function (user) {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                var query = {
+                    "Pequena": req.body.item.PrecoMarmitex.Pequena,
+                    "Grande": req.body.item.PrecoMarmitex.Grande
+                };
+                Mongo_1.Mongo.Edit(MongoCollections_1.Collections.PrecoMarmitex.NomeID, req.body.item.PrecoMarmitex._id, query).then(function (x) {
+                    redisConfig_1.redisConfig.flushAll();
+                    res.send(x);
+                });
+            }
         });
     }
     catch (err) {
@@ -96,12 +117,17 @@ app.put(Routes_1.Rotas.PrecoMarmitex, function (req, res) {
 });
 app.put(Routes_1.Rotas.Complemento, function (req, res) {
     try {
-        var query_1 = { "Nome": req.body.Complemento.Nome, "Tipo": req.body.Complemento.Tipo, "Preco": req.body.Complemento.Preco };
-        console.log(query_1);
-        Mongo_1.Mongo.Filtrar(MongoCollections_1.Collections.Complemento.NomeID, { "_id": req.body.Complemento._id }).then(function (result) {
-            Mongo_1.Mongo.EditarPorAtributo(MongoCollections_1.Collections.Complemento.NomeID, { "Nome": req.body.Complemento.Nome }, query_1);
-            res.send({ sucesso: req.body });
-            redisConfig_1.redisConfig.flushAll();
+        usuarios_service_1.UsuarioService.getByToken(req.body.token).then(function (user) {
+            console.log(user);
+            if (user[0].Tipo == 2) {
+                var query_1 = { "Nome": req.body.item.Complemento.Nome, "Tipo": req.body.item.Complemento.Tipo, "Preco": req.body.item.Complemento.Preco };
+                console.log(query_1);
+                Mongo_1.Mongo.Filtrar(MongoCollections_1.Collections.Complemento.NomeID, { "_id": req.body.item.Complemento._id }).then(function (result) {
+                    Mongo_1.Mongo.EditarPorAtributo(MongoCollections_1.Collections.Complemento.NomeID, { "Nome": req.body.item.Complemento.Nome }, query_1);
+                    res.send({ sucesso: req.body });
+                    redisConfig_1.redisConfig.flushAll();
+                });
+            }
         });
     }
     catch (err) {
